+++ local_planner/include/local_planner/avoidance_output.h
