+++ local_planner/include/local_planner/common.h
