+++ local_planner/include/local_planner/waypoint_generator.h
