+++ local_planner/src/nodes/box.cpp
@@ -1,4 +1,4 @@
-#include "box.h"
+#include "local_planner/box.h"
 
 namespace avoidance {
