+++ local_planner/src/nodes/common.cpp
@@ -1,4 +1,4 @@
-#include "common.h"
+#include "local_planner/common.h"
 
 #include <cmath>
 #include <fstream>
