+++ local_planner/src/nodes/local_planner_node.cpp
@@ -1,10 +1,8 @@
-#include "local_planner_node.h"
-
-#include "common.h"
-#include "local_planner.h"
-#include "planner_functions.h"
-#include "tree_node.h"
-#include "waypoint_generator.h"
+#include "local_planner/local_planner_node.h"
+#include "local_planner/local_planner.h"
+#include "local_planner/planner_functions.h"
+#include "local_planner/tree_node.h"
+#include "local_planner/waypoint_generator.h"
 
 #include <boost/algorithm/string.hpp>
