+++ local_planner/src/nodes/star_planner.cpp
@@ -1,8 +1,7 @@
-#include "star_planner.h"
-
-#include "common.h"
-#include "planner_functions.h"
-#include "tree_node.h"
+#include "local_planner/star_planner.h"
+#include "local_planner/common.h"
+#include "local_planner/planner_functions.h"
+#include "local_planner/tree_node.h"
 
 #include <ros/console.h>
